Clean up and moving stuff around
