+++ projects/psvr2_openvr_driver_ex/libpad_hooks.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include "rolling_percentile.h"
 #include <mutex>
 
 namespace psvr2_toolkit {
