+++ projects/psvr2_openvr_driver_ex/psvr2_openvr_driver_ex.vcxproj
@@ -210,7 +210,6 @@
     <ClInclude Include="math_helpers.h" />
     <ClInclude Include="minhook\include\MinHook.h" />
     <ClInclude Include="pad_trigger_effect.h" />
-    <ClInclude Include="rolling_percentile.h" />
     <ClInclude Include="sense_controller.h" />
     <ClInclude Include="sense_crc.h" />
     <ClInclude Include="psvr2_openvr_driver\openvr\headers\openvr_driver.h" />
