+++ projects/psvr2_openvr_driver_ex/usb_thread_hooks.cpp
@@ -5,6 +5,8 @@
 #include "vr_settings.h"
 
 namespace psvr2_toolkit {
+  std::mutex ldPayloadMutex;
+  LDPayload currentLDPayload;
 
   int (*CaesarUsbThread__report)(void *thisptr, bool bIsSet, uint16_t reportId, void *buffer, uint16_t length, uint16_t value, uint16_t index, uint16_t subcmd);
 
@@ -15,11 +17,11 @@ namespace psvr2_toolkit {
     return result;
   }
 
-  LDPayload currentLDPayload;
-
   uint64_t(*CaesarUsbThreadLeddet__poll)(void* thisptr) = nullptr;
   uint64_t CaesarUsbThreadLeddet__pollHook(void* thisptr) {
     uint64_t result = CaesarUsbThreadLeddet__poll(thisptr);
+
+    std::scoped_lock<std::mutex> lock(ldPayloadMutex);
 
 	  memcpy(&currentLDPayload, reinterpret_cast<uint8_t*>(thisptr) + 0x230, sizeof(LDPayload));
